+++ src/Commands/Enums/GroupMemberType.cs
@@ -1,0 +1,23 @@
+namespace PnP.PowerShell.Commands.Enums
+{
+	/// <summary>
+	/// Specifies the member type for a SharePoint Online geo administrator.
+	/// </summary>
+	public enum GroupMemberType
+	{
+		/// <summary>
+		/// The member type is unknown.
+		/// </summary>
+		Unknown = 0,
+
+		/// <summary>
+		/// The member is a user.
+		/// </summary>
+		User = 1,
+
+		/// <summary>
+		/// The member is a group.
+		/// </summary>
+		Group = 2
+	}
+}