+++ src/Commands/Utilities/MultiGeo/MultiGeoRestApiClient.cs
@@ -31,17 +31,28 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 		private const string TenantRenameJobsPathToCancelAJob = "TenantRenameJobs/Cancel";
 		private const string GeoMoveCompatibilityChecksMinimumApiVersion = "1.3.6";
 		private const string GeoMoveCompatibilityChecksPath = "GeoMoveCompatibilityChecks";
+		private const string GeoAdministratorsMinimumApiVersion = "1.2-beta";
+		private const string GeoAdministratorsPath = "GeoAdministrators";
 		private const string GeoExperienceMinimumApiVersion = "1.3.7";
 		private const string GeoExperiencePath = "GeoExperience";
 		private const string UpdateGeoExperienceModePath = "GeoExperience/UpgradeToSPOMode";
 		private const string UpdateAllInstancesExperienceModePath = "GeoExperience/UpgradeAllInstancesToSPOMode";
 		private const string AllowedDataLocationsApiVersion = "1.3.11";
 		private const string AllowedDataLocationsPath = "AllowedDataLocations";
+		private const string AllowedDataLocationByLocationPath = "AllowedDataLocations(location='{0}')";
+		private const string GeoAdministratorsByLoginNameMaximumApiVersion = "1.3.8";
+		private const string GeoAdministratorsByPrincipalMinimumApiVersion = "1.3.9";
+		private const string GeoAdministratorsByLoginNamePath = "GeoAdministrators(loginName='{0}')";
+		private const string GeoAdministratorsByLoginNameAndTypePath = "GeoAdministrators/GetByLoginNameAndType(loginName='{0}', memberType={1:D})";
+		private const string GeoAdministratorsByObjectIdPath = "GeoAdministrators/GetByObjectId(guid'{0:D}')";
 		private const string StorageQuotasMinimumApiVersion = "1.3.1";
 		private const string StorageQuotasPath = "StorageQuotas";
 		private const string StorageQuotaByLocationPath = "StorageQuotas(geoLocation='{0}')";
 		private const string MultiGeoApiVersionsPath = "MultiGeoApiVersions";
+		private const string DeleteVerbString = "DELETE";
 		private const string PatchVerbString = "PATCH";
+		private const string UserPersonalSiteLocationMinimumApiVersion = "1.0";
+		private const string UserPersonalSiteLocationPath = "UserPersonalSiteLocation('{0}')";
 		private const string UserMoveJobsMinimumApiVersion = "1.0";
 		private const string UserMoveJobsByMoveIdMinimumApiVersion = "1.2.2";
 		private const string UserMoveJobsReportMinimumApiVersion = "1.3.2";
@@ -64,6 +75,8 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 		private const int SiteMoveRunsWorkflow2013ErrorCode = -113;
 		private const int SiteMoveRequiresForceErrorCode = -116;
 		private const int SiteMoveContainsBcsErrorCode = -139;
+		private const int GeoAdministratorUserMemberType = 1;
+		private const int GeoAdministratorGroupMemberType = 2;
 		private const int MaximumPagination = 10;
 		private const int ApiVersionCacheValidTimeInHours = 1;
 		private static readonly TimeSpan CreateTenantRenameJobTimeout = TimeSpan.FromSeconds(300);
@@ -158,6 +171,11 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 			return GetFeed<GeoMoveTenantCompatibilityCheck>(GeoMoveCompatibilityChecksPath, GetCurrentApiVersion(GeoMoveCompatibilityChecksMinimumApiVersion));
 		}
 
+		internal GeoAdministratorCollection GetGeoAdministrators()
+		{
+			return Get<GeoAdministratorCollection>(GeoAdministratorsPath, GetGeoAdministratorsApiVersion());
+		}
+
 		internal MultiGeoExperience GetGeoExperience()
 		{
 			return Get<MultiGeoExperience>(GeoExperiencePath, GetGeoExperienceApiVersion());
@@ -166,6 +184,26 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 		internal IEnumerable<MultiGeoCompanyAllowedDataLocation> GetAllowedDataLocations()
 		{
 			return GetFeed<MultiGeoCompanyAllowedDataLocation>(AllowedDataLocationsPath, AllowedDataLocationsApiVersion);
+		}
+
+		internal GeoAdministrator AddGeoAdministrator(GeoAdministratorEntityData geoAdministrator)
+		{
+			if (geoAdministrator == null)
+			{
+				throw new ArgumentNullException(nameof(geoAdministrator));
+			}
+
+			return Post<GeoAdministrator>(GeoAdministratorsPath, geoAdministrator, apiVersion: GetGeoAdministratorsApiVersion());
+		}
+
+		internal void EnsureGeoAdministratorObjectIdSupported()
+		{
+			GetGeoAdministratorsByPrincipalApiVersion();
+		}
+
+		internal void EnsureGeoAdministratorSupported()
+		{
+			GetGeoAdministratorsApiVersion();
 		}
 
 		internal void UpgradeGeoExperience(bool allInstances)
@@ -195,6 +233,18 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 			PostWithoutResponse(AllowedDataLocationsPath, allowedDataLocation, apiVersion);
 		}
 
+		internal void RemoveAllowedDataLocation(string location)
+		{
+			var apiVersion = GetCurrentApiVersion();
+			if (!IsSupportedApiVersion(apiVersion, AllowedDataLocationsApiVersion))
+			{
+				throw new NotSupportedException(string.Format(CultureInfo.InvariantCulture, CommandResources.CrossGeoInvalidVersion, GetApplicationVersion()));
+			}
+
+			var path = string.Format(CultureInfo.InvariantCulture, AllowedDataLocationByLocationPath, ProcessSpecialChars(location));
+			PostWithMethodOverrideEmptyBody(path, DeleteVerbString, apiVersion);
+		}
+
 		internal IEnumerable<StorageQuota> GetStorageQuotas()
 		{
 			return GetFeed<StorageQuota>(StorageQuotasPath, GetStorageQuotasApiVersion());
@@ -212,6 +262,13 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 			var apiVersion = GetCurrentApiVersion(UserMoveJobsMinimumApiVersion);
 			var path = string.Format(CultureInfo.InvariantCulture, UserMoveJobPathByUpn, ProcessSpecialChars(userPrincipalName));
 			return Get<UserAndContentMoveState>(path, apiVersion);
+		}
+
+		internal UserPersonalSiteLocation GetUserPersonalSiteLocation(string userPrincipalName)
+		{
+			var apiVersion = GetCurrentApiVersion(UserPersonalSiteLocationMinimumApiVersion);
+			var path = string.Format(CultureInfo.InvariantCulture, UserPersonalSiteLocationPath, ProcessSpecialChars(userPrincipalName));
+			return Get<UserPersonalSiteLocation>(path, apiVersion);
 		}
 
 		internal UserAndContentMoveState GetUserAndContentMoveState(Guid odbMoveId)
@@ -333,6 +390,23 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 			PostWithMethodOverride(path, quota, PatchVerbString, apiVersion);
 		}
 
+		internal void RemoveGeoAdministrator(string loginName, bool isGroup)
+		{
+			var apiVersion = GetGeoAdministratorsApiVersion();
+			var path = IsSupportedApiVersionRange(apiVersion, GeoAdministratorsMinimumApiVersion, GeoAdministratorsByLoginNameMaximumApiVersion)
+				? string.Format(CultureInfo.InvariantCulture, GeoAdministratorsByLoginNamePath, ProcessSpecialChars(loginName))
+				: string.Format(CultureInfo.InvariantCulture, GeoAdministratorsByLoginNameAndTypePath, ProcessSpecialChars(loginName), isGroup ? GeoAdministratorGroupMemberType : GeoAdministratorUserMemberType);
+
+			PostWithMethodOverrideEmptyBody(path, DeleteVerbString, apiVersion);
+		}
+
+		internal void RemoveGeoAdministrator(Guid objectId)
+		{
+			var apiVersion = GetGeoAdministratorsByPrincipalApiVersion();
+			var path = string.Format(CultureInfo.InvariantCulture, GeoAdministratorsByObjectIdPath, objectId);
+			PostWithMethodOverrideEmptyBody(path, DeleteVerbString, apiVersion);
+		}
+
 		internal void CancelUserMoveJob(string userPrincipalName)
 		{
 			var apiVersion = GetCurrentApiVersion(UserMoveJobsMinimumApiVersion);
@@ -425,6 +499,21 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 			Send(() =>
 			{
 				var request = CreateRequest(HttpMethod.Post, path, apiVersion, jsonPayload);
+				request.Headers.TryAddWithoutValidation("X-HTTP-Method", methodOverride);
+				if (request.Content != null)
+				{
+					request.Content.Headers.ContentType = MediaTypeHeaderValue.Parse("application/json;charset=UTF-8");
+				}
+
+				return request;
+			}, timeout: null, allowRetries: false);
+		}
+
+		private void PostWithMethodOverrideEmptyBody(string path, string methodOverride, string apiVersion)
+		{
+			Send(() =>
+			{
+				var request = CreateRequest(HttpMethod.Post, path, apiVersion, string.Empty);
 				request.Headers.TryAddWithoutValidation("X-HTTP-Method", methodOverride);
 				if (request.Content != null)
 				{
@@ -528,10 +617,42 @@ namespace PnP.PowerShell.Commands.Utilities.MultiGeo
 			return apiVersionIndex >= 0 && minimumApiVersionIndex >= 0 && apiVersionIndex <= minimumApiVersionIndex;
 		}
 
+		private static bool IsSupportedApiVersionRange(string apiVersion, string minimumApiVersion, string maximumApiVersion)
+		{
+			var apiVersionIndex = Array.IndexOf(ClientSupportedApiVersions, apiVersion);
+			var minimumApiVersionIndex = Array.IndexOf(ClientSupportedApiVersions, minimumApiVersion);
+			var maximumApiVersionIndex = Array.IndexOf(ClientSupportedApiVersions, maximumApiVersion);
+			return apiVersionIndex >= 0 && minimumApiVersionIndex >= 0 && maximumApiVersionIndex >= 0
+				&& apiVersionIndex <= minimumApiVersionIndex
+				&& apiVersionIndex >= maximumApiVersionIndex;
+		}
+
 		private string GetStorageQuotasApiVersion()
 		{
 			var apiVersion = GetCurrentApiVersion();
 			if (!IsSupportedApiVersion(apiVersion, StorageQuotasMinimumApiVersion))
+			{
+				throw new NotSupportedException(string.Format(CultureInfo.InvariantCulture, CommandResources.CrossGeoInvalidVersion, GetApplicationVersion()));
+			}
+
+			return apiVersion;
+		}
+
+		private string GetGeoAdministratorsApiVersion()
+		{
+			var apiVersion = GetCurrentApiVersion();
+			if (!IsSupportedApiVersion(apiVersion, GeoAdministratorsMinimumApiVersion))
+			{
+				throw new NotSupportedException(string.Format(CultureInfo.InvariantCulture, CommandResources.CrossGeoInvalidVersion, GetApplicationVersion()));
+			}
+
+			return apiVersion;
+		}
+
+		private string GetGeoAdministratorsByPrincipalApiVersion()
+		{
+			var apiVersion = GetCurrentApiVersion();
+			if (!IsSupportedApiVersion(apiVersion, GeoAdministratorsByPrincipalMinimumApiVersion))
 			{
 				throw new NotSupportedException(string.Format(CultureInfo.InvariantCulture, CommandResources.CrossGeoInvalidVersion, GetApplicationVersion()));
 			}